Sellaroshop/requirements.txt
asgiref==3.10.0
Django==5.2.7
pillow==12.0.0
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.3.1
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.13.0
requests==2.31.0
python-decouple==3.8
